# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce API",
        "description": "Dokumentasi API untuk aplikasi e-commerce",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Server utama"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Mendaftarkan pengguna baru",
                "description": "Endpoint untuk mendaftarkan akun baru menggunakan Sanctum",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registrasi berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login pengguna",
                "description": "Endpoint untuk login dan mendapatkan token menggunakan Sanctum",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout pengguna",
                "description": "Menghapus token pengguna untuk logout",
                "operationId": "logoutUser",
                "responses": {
                    "200": {
                        "description": "Logout berhasil"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/invoices/{order_id}": {
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Generate invoice dari order yang telah selesai",
                "operationId": "c0b2ce8b85155589e21a52e1a5c25b37",
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "path",
                        "description": "ID Order yang akan dibuatkan invoice",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "discount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10.5
                                    },
                                    "custom_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150000
                                    },
                                    "invoice_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-02-26"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Invoice berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invoice berhasil dibuat"
                                        },
                                        "invoice": {
                                            "$ref": "#/components/schemas/Invoice"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Order belum selesai atau total harga negatif"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/invoices": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Ambil daftar semua invoice",
                "operationId": "0e8c171ac297860339fe771f1e166afd",
                "responses": {
                    "200": {
                        "description": "Daftar invoice berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "invoices": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Invoice"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "RegisterRequest": {
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Fadel Razsiah"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "fadelrazsiah@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "password123"
                    }
                },
                "type": "object"
            },
            "LoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "fadelrazsiah@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "password123"
                    }
                },
                "type": "object"
            },
            "AuthResponse": {
                "properties": {
                    "token": {
                        "type": "string",
                        "example": "1|aLongGeneratedSanctumToken..."
                    }
                },
                "type": "object"
            },
...
            "Invoice": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "order_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 100000
                    },
                    "invoice_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-02-26"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-02-26T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-02-26T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    ]
}